Changed the creation size from 500 to 3000 since most of the small system of 300 classes will have a size which is near to 3000.
The Overload value has been changed from 10000 to 50000 so that it will reduce the
number of writting and which will also the size of the EIFGEN since each time we
create a new file to replace the old file, we forget the old file and at the end it
is not deleted.

So by changing this value, we will reduce the number of forgotten file until we
find a better implementation.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@11291 8089f293-4706-0410-a29e-feb5c42a2edf