Cosmetics
Add initialization part and ending part to the degree_minus_5 for the polymorphic table.

We now process degree_minus_4 in the way of the topological sort.
For the Dead Code Removal we are not keeping all the base class. Just the one
we need.

Removed the generation of the conformance table in final mode.

Changed the way how to generate the polymorphic table by using the `used' table
for EIFFEL_HISTORY.

Changed the way how to generate the skeleton file, we are generating directly the
offsets and no more indirection on those offset through the polymorphic table.

Improve the speed of the generation of the dispose table, we don't use anymore
the `dispose_feature' from CLASS_C but the polymorphic table associated with
the routine id of `dispose' from MEMORY.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@11842 8089f293-4706-0410-a29e-feb5c42a2edf