API/interface/e_feature.e eiffel/AST/clickable/class_type_as.e eiffel/AST/expressions/binary_as.e eiffel/AST/expressions/un_strip_as.e eiffel/AST/expressions/unary_as.e eiffel/AST/skeleton/access_address_as.e eiffel/AST/skeleton/access_feat_as.e eiffel/AST/skeleton/bits_symbol_as.e eiffel/AST/skeleton/id_as.e eiffel/AST/skeleton/like_id_as.e eiffel/AST/skeleton/precursor_as.e eiffel/byte_code/feature_b.e eiffel/compiler/array_optimizer.e eiffel/compiler/class_dependance.e eiffel/compiler/depend_unit.e eiffel/compiler/feat_iterator.e eiffel/compiler/feature_dependance.e eiffel/compiler/feature_table.e eiffel/compiler/remover.e eiffel/compiler/select_table.e eiffel/generation/attr_entry.e eiffel/generation/entry.e eiffel/generation/poly_table.e eiffel/generation/rout_entry.e eiffel/generation/special_entry.e eiffel/inheritance/inherit_table.e eiffel/interface/attribute_i.e eiffel/interface/class_c.e eiffel/interface/external_i.e eiffel/interface/feature_i.e eiffel/interface/system_i.e eiffel/server/depend_server.e
New dead code removal: the old dead code removal was going through the
FEATURE_I of the system in a non-linear way which took a very long
time since the FEATURE_I are not very fast to be retrieved (We were
retrieving them through the FEATURE_TABLE). Now, the
algorythm is basically the same, we go through the whole system marking
the features alive and then their dependances, but we transfered the
usefull information for the algorithm in the dependances themselves
making it a more compact information to retrieve. Thus we had also to
update all the calls to the feature dependances in the system.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@11912 8089f293-4706-0410-a29e-feb5c42a2edf