Improved the C generation by specifying a complete signature for features which were not generated in the Current file. So now instead of generating something like: 	extern EIF_REFERECNE Easd34s (); we are generating the correct signature: 	extern EIF_REFERENCE Easd34s (EIF_REFERENCE, EIF_INTEGER, EIF_REFERENCE);
This change was needed for the CPP compilation using `big_file' done by `quick_finalize'.
Also, it shows that we are generating correct ANSI C code.

This change has not been done at the level of the E1 directory, since the tables
do not require any precise signature, that's maybe something we want to do later.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@15135 8089f293-4706-0410-a29e-feb5c42a2edf