Used EIF_REFERENCE instead of `char *'.
Fixed a very old bug in the debugger which was easy to fix thanksful to the same
kind of bug found in `rout_obj.c'. Basically we were creating two eiffel strings
without protecting the references. As a result, the first string could have moved
after the creation of the second one and most of the time is was crashing EiffelBench.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@17024 8089f293-4706-0410-a29e-feb5c42a2edf