By fixing the previous bug when a redefinition which was returning `like Current' we introduced a new bug which was making the following to crash class	A inherit	LINKED_LIST [INTEGER] redefine item end feature	item: INTEGER is do Result := Precursor end end
The reason is that item is defined of type G and we need to solved the type of
Precursor therefore in the context LINKED_LIST[INTEGER] and not LINKED_LIST[G].

To fix that, we check that the declared type of `Precursor' does not return a
formal_type, if it does we evaluate it in the context of the inheritance, otherwise
we evaluate it in the context of the parent.

Also removed some line of codes which were not used because Precursor is never called
in the case of a "." expression, `last_type' is always a valid type, i.e. no anchored
no generic parameter.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@18139 8089f293-4706-0410-a29e-feb5c42a2edf