Fixed a warning by adding a return value to `ora_next_row'.
Fixed a bug in functions calling in a for loop `ora_terminate_order'
because they were not using the current number of descriptor, but the
maximum number of descriptor which could conduce to strange problems.

Fixed a bug in `ora_get_date_data' where if there is a NULL value, we return
the default date `01/01/0000 00:00:00'.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@18528 8089f293-4706-0410-a29e-feb5c42a2edf