Cosmetics on code and removed useless variable initialization.
Fixed in `eif_typeof_array' the code so that it works fine with Borland C++ 5.5.1. Even
though I'm convinced the code is correct, it does not work under Borland. The rationale
is given here as an explanation of what is happening and the appropriate fix:

In `eif_typeof_array' we have the following piece of code:
int16 typearr[4];
typearr [0] = -1;
...
typearr [3] = TERMINATOR;

eif_compound_id (..., typearr, ...);

When using a precompiled library and a program that have a creation routine that expects
an array of string as argument, the run-time crashed in the computation of `eif_compound_id'
down in `eif_id_of'.

Using the debugger, I found out the following: before doing the line typerr [0],
suppose the array is initialized to the following values: -1, -2 ,-3 ,-4
(random values). After one recursion in `eif_id_of' instead of having **intab equal
to `typearr [3]' as I expected, I got -3, the value given when the `typearr'
array was allocated.

Changing the `typearr' initialization by:
int16 *typearr;
typearr = (int16 *) malloc (4 * sizeof (int16);
....
eif_compound_id (..., typearr, ...);
free (typearr);

worked out just all the time.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@26844 8089f293-4706-0410-a29e-feb5c42a2edf