Reverted back to version 1.61 because the assumption made that no EIF_GEN_DER structure was needed for non-generic classes is not completely True, because if you use the `par_info' structure you need to do some complicated computation that are performed when the EIF_GEN_DER structure is created.
Fixed bug again in a different manner in generic conformance where an assignment
attempt could affect the result of a generic creation, it corresponds to the EiffelWeasel
entry "assign-attempt-affects-source" (exec121). (Big fix is at line 2020)

Since we fixed `update.c' so that the `parents' field of a `par_info' structure
cannot be NULL, I removed the check against NULL and removed the `nulltab' array
that is not needed anymore.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@27859 8089f293-4706-0410-a29e-feb5c42a2edf