Fixed bug in precursor with the following system: class C [G] feature 	item: G is do end 	f: LIST [C [G]] is do end end -- class C
class
	D
inherit
	C [STRING] redefine f end
feature
	f: LIST [C [like item]] is
		do
			Result := Precursor
		end
end -- class D

It was crashing because the signature of `Precursor' in D was still evaluated
as if it was in C [G], but it should have been evaluated in the context of
C [STRING]. So now we instantiate the found `feature_i' from the parent class
with the parent class instantiation in the current class.

Optimized to go directly to the FEATURE_I instead of going through the E_FEATURE
which are very nasty.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@29857 8089f293-4706-0410-a29e-feb5c42a2edf