Fixed `internal_select_item' and `internal_deselect_item'. Although we were setting the item to be selected, we were not setting the `Lvis_focused' flag. This showed up as a bug in the following case:
Set the focus to a list.
Call list_item.enable_select
press the up or down arrow to move the focus within the list.

No matter which list_item was selected, the focus was still at position
-1. This means that you could only use the down arrow, and it would then
select the first item in the list. This was not nice behaviour.

Also, when using `internal_deselect_item', the list was still focused to
the previously selected item, although the item was no longer selected.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@30088 8089f293-4706-0410-a29e-feb5c42a2edf