Fixed a bug in code generation of reverse assignment on attribute of a class. If you had: 	a: A 	a ?= xxx where `a' is an attribute of the current class. If in a descendant class you redefine `a' to have a different type, the code was not taking into account the new type of `a' but the original one, i.e. `A'.
To fix this problem, we introduced a new `info' attribute to REVERSE_B and made
REVERSE_BL inherit from REVERSE_B in order to get this new attribute. Also moved
the code `make_reverse_code' from ACCESS_B into REVERSE_B as it was the incorrect place.

Now in `make_byte_code' and `generate_regular_assignment of REVERSE_B/_BL class
we take into account `info' to find out what is the real type of `a' in the
descendant class in order to have a good behavior of the assignment attempt.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@31037 8089f293-4706-0410-a29e-feb5c42a2edf