The following changes have been made in `unconnect_radio_gouping' to fix bugs:
We now check that there was an original radio button selected, only if
the radio group is not empty.

We now only perform selection at the end if there was originally a
radio button selected.

If the original selected radio button was in `Current', then we only
select `peer' if it has a radio button.

These changes are mainly for the cases when containers are linked, but
have no radio buttons, which was overlooked originally.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@32081 8089f293-4706-0410-a29e-feb5c42a2edf