Changed access level of some features which should not have been accessible to any other class by SYSTEM_CLUSTERS. Added more preconditions to a number of features to ensure absolute safety when calling.
Fixed bug in expanded_path where function would go into an infinite loop when
trying to evaluate $VAR1$VAR2 as path, or when $VAR was not found.
Function previous could not evaluate $VAR1$VAR2 as it was trying to
find env var 'VAR1$VAR2' becuase it was not looking for $ as var seperator.
Now works.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@33658 8089f293-4706-0410-a29e-feb5c42a2edf