Added comment to `insertion_position' and fixed bug in this feature. To reproduce the bug that was due to the implementation of `insertion_position', perform the following:
set up a tool bar containing a number of buttons and seperators.
Ensure that the tool bar and its contents are dockable.
Start to transport one of the buttons that has a seperator immediately to
its left
Release the transport while above the seperator.


This will crash the system, as `insertion_position' was returning 0, as
the old implementation was performing a reverse assignment on a tool bar
button, and the type was a seperator.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@36218 8089f293-4706-0410-a29e-feb5c42a2edf