Now inherits GB_POST_OBJECT_EXPANDER to allow the expanded state of the layout constructor to be saved and restored from the save file.
The reason that we must introduce a new class for this, is that GB_OBJECT
is the only class that knows how to read the XML file directly, but other
classes must perform the loading and rebvuilding. Therefore we need a
common class which may be used by these different classes.

`generate_xml' and `modify_from_xml' have been modified to allow this
restoration.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@36446 8089f293-4706-0410-a29e-feb5c42a2edf