Corrected bug in `set_item_position_and_span' which would cause a crash if you performed the following:
create a table
resize to say 6x6
insert an item, and then do the following:
table.set_item_position_and_span (item, 5, 1, 2, 6)
Table.set_item_position_and_span (item, 6, 1, 1, 6)

The version of `set_item_position_and_span' would still use the internal
span of the table item, during its implementation which would cause a crash
as it would attempt to modify cells that did not fall within the boundaries
of the table.

To solve this, we now redefine `set_item_position_and_span' within
EV_TABLE_IMP and firstly update the internal representation of the table
child internally. Then when we start moving the item, the spans are
correct and we no longer attempt to modify cells that do not exist.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@36467 8089f293-4706-0410-a29e-feb5c42a2edf