Fixed bug in `on_wm_command' which could be reproduced as follows:
create and display a dialog modally to a window, with a default push button
that will launch a second modal dialog relative to the original dialog.
Within the second dialog, insert a text field, whose `return_actions'
will destroy the second dialog.

Before this fix, the second dialog will be redisplayed every time that
return is pressed within the text field.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@37250 8089f293-4706-0410-a29e-feb5c42a2edf