New code generation for agents which enables us to have more efficient agent calls (look at CVS log of version 1.4 of ROUTINE class for details on speed improvements and required library changes).
We are now only generating a tuple which will be a placeholder for all
open and closed operands of routine, and an array which contains all
open position in above mentioned tuple. That way we only need to take
care of inserting open operands at the right position in the tuple just
before the call.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@37459 8089f293-4706-0410-a29e-feb5c42a2edf