Refactored code for generation scavenging so that it is more clear what the value of `gen_scavenge' means. That is to say we have now 3 states instead of 4. It is either `disabled', `enabled' and `enabled but stopped'. When it is disabled, sometime the GC will try to activate it again, before it was done by setting the GS_SET flag and let `malloc_from_zone' recreate the savenge zones. Now, instead of putting a flag, the GC tries to create those scavenge zones. If it succeeds then of course `malloc_from_zone' will be called.
By doing so, we reduced the size of `malloc_from_zone' which can be inlined
(at least on Windows platforms using VC++). We mostly removed code that tested
if allocation in scavenge zone was allowed and if so and no scavenge zone
were available we were forcing a creation. This, by the way, was a major issue
for multithreaded application that we don't have to worry anymore. This
improvement in the size of `malloc_from_zone' speeds up our fast memory
allocation. On Unix, we noticed a gain of 5s from 1104 to 1099s of compile time.

As a consequence of this change `mem_speed' has been changed to force creation
of the scavenge zone if needed. Also the creation of scavenge zone is now
taken care in `eif_alloc_init' in non-multithreaded application and in
`eif_thr_init_root' for multithreaded application (of course `cc_for_speed'
needs to be set to 1 otherwise they are not needed).

By doing a review of callers of `malloc_from_zone' to ensure its is called
properly (i.e. gen_scavenge == GS_ON), I've noticed that `spmalloc' was not
nicely coded. I've refactored the code to look more like `emalloc_size'. First
we try to allocate in the scavenge zone if possible, then in the free list.
Note that if both allocation failed we don't try to do more like we do
in `emalloc_size'. Maybe that's something we ought to fix in the future.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@38599 8089f293-4706-0410-a29e-feb5c42a2edf