New thread safe memory allocation/GC collection. In addition of that we made a major speed improvements of the allocation/GC collection mechanism which provides a gain of roughly 15% in speed. You will find below the changes we made to achieve that. Tested on the compiler we went from 1102s to 934s and we even removed our calls to `full_collect' and `full_coalesc' we added a long time ago to improve runtime performance (when we kept them the amelioration was of only 945s, so not having them and let the runtime decides for us is a better strategy).
Now when we allocate, we always try to allocate in the scavenge zone, if it
fails we either do a small collect or we do an automatic one if necessary and
we try again.
If it fails, then we try to find an empty spot in the free list. If it fails
we perform an automic collection if necessary, and we retry again.
If it fails we go to `malloc_free_list' which depending on the value of `cc_for_speed'
will either allocate a new chunk of memory or launch a full collection. At
this point we should have some memory. This is by doing this decomposition
in the memory allocation that we were able to made it thread safe. A thumb of
rule is that the thread safe allocation routines cannot call the GC, only the
GC can call them (important to avoid dead-lock).

As you may notice we try delay the full collection as much as possible as
it is costly. The second improvement is actually on the full collection.
Before `plsc_per' was a fixed value. Now it is not fixed anymore it is a variable
value. The idea is that each time we do a full collection, we analyze if
we have freed a lot memory or almost nothing. If we did a lot, then it means
that we should decrease the current value of `plsc_per' to have less to free
each time so that we don't allocate to much from the kernel. If we are in
between (not a lot, not nothing) we don't change it. If we free almost nothing
then we increase the value of `plsc_per' so that next time we hope to free more.
From 20 we go to 100 for `plsc_per' when compiling the batch compiler showing
that it made sense to have it variable rather than fixed.

One more thing we noticed that sometime with this new configuration we were
using too much memory, We found out that the reason was due to the moving
of objects from the scavenge zone to outside. Because in most cases we
did not have enough space in the chunks of memory. Therefore we were causing
multiple calls to `allocate_from_core' which was allocating memory that
at the end was collected next time we called a full collection. To avoid this
situation, we have forced for the next call to `acollect' a full collection by
incrementing `force_pslc' when this happens.

Renamed `gmalloc' into `malloc_from_eiffel_list_no_gc'.
Merged `set_up' and `set_up_chunk' into a single routine as we don't make the difference
  anymore between a C chunk and an Eiffel one.
Factorized `allocate_free_list' to avoid code duplication of critical loop.
Size of C chunks is not paded by `eif_chunk_size' anymore, but by PAGESIZE_VALUE
  instead.
Removed `xfreechunk' in favor of `eif_rt_xfree' as they had the same implementation.
Added `trigger_gc_cyle' and `trigger_smart_gc_cycle' which are entry point
  to collection routines.
Speed up `malloc_from_zone' by reducing its size to favorize inlining, we do
  not try to recover when it fails, it is up to the caller to decide what to do.
Added `att_to_stack' a generic routine to avoid code duplication.

Added full documentation of routines for `malloc.c'.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@38702 8089f293-4706-0410-a29e-feb5c42a2edf