Fixed `remove_docking_areas' so that it fixed the following bug which crashed the system: Minimize the first two items in the area. Drag the second, so that it is now the first. Restore both minimized items.
update_for_holder_position_change has been modified so that it takes the
holder, and not the original position, as the original position may
be gathered from the holder, which helps support the case where you
drag back from an external dialog.

Fixed bug in `initialize_docking_areas' which was not taking into the
account when an item awas external, and being docked back in.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@39062 8089f293-4706-0410-a29e-feb5c42a2edf