`format_region' now ensures that the caret position does not change, which was causing the postcondition to fail.
`buffered_append' now correctly uses the background color.

`buffered_format' now correctly uses the background color. Also fixed buffering
in this mode, as before, calling `flush_buffer' did not flush it to the
correct positions. Fixed another bug that caused the formatting for any of the
text that was not included within calls to `buffered_append' to become messed up
with incorrect formatting. To fix this, we now only stream from the lower and upper limits of the calls to `bufered_format', replacing the existing text
and leaving the remaining text outside these boundaries untouched.
The default format applied in the range of formatting to be replaced by
`buffered_format' now uses the correct background color of the control.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@42022 8089f293-4706-0410-a29e-feb5c42a2edf