Fixed bug in `set_layout_item' which caused EiffelBuild to crash if you performed the following:
Reset an object, and then pick its representation from the layout selector
and try tro build somewhere within the interface.

When we reset an object, we keep the same layout item but update its object
reference. Unfortunately, the `pebble_function' associated to the layout item
is still the old one that has originally set, causing a pick to generate the
old object. Now, we reset the pebble function to ensure that when picking we get
the new object.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@46967 8089f293-4706-0410-a29e-feb5c42a2edf