Undid previous change which broke handling of dialogs from the drop actions of a pick and dropable widget. We now call the new application level query `drop_actions_executing' to prevent a pick starting from a drop which is a new fix which performs the same function as the last commit, without the buggy side effect.
The bug that appeared with the last fix is as follows:

From the drop actions of a pick and dropable target, display a dialog modaly
and the cursor handling of the dialog is incorrect.

The bug was due to the fact that the we calling `disable_default_processing'
in EV_WINDOW_IMP to prevent windows raising as required for pick and drop.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@47384 8089f293-4706-0410-a29e-feb5c42a2edf