Added once optimizations to post_degree_5_execute
Also minor optimization for passing 'old_info' to fill_parents, all that is needed is a call to 'server_item' as this will return Void if not present in the ast_server.  This removes the need for an extra call to 'has' in the previous code when the item was present, plus removes an unnecessary local variable usage.

git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@72048 8089f293-4706-0410-a29e-feb5c42a2edf