Now correctly determining whether a call that needs renaming gets detected correctly as this information is only available incrementality from the AS object itself.
Fixed issue with renaming in a single branch where the name changes but routine id does not (ie: single inheritance) and this routine is then unselected and replicated in a descendent but the AST still points to the original name which may now be the selected branch, and therefore wrong routine.  Thanks to Ted for finding this issue.

git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@75430 8089f293-4706-0410-a29e-feb5c42a2edf