Fixed bug#10736 that the following command on Solaris caused runtime panic:  sh  cd /  cat /dev/null | ec
The reason it only happens on Bourne shell (sh) is that it behaves differently than other shells when it sets up a pipeline. Other shells create two processes that are a child of the parent - i.e. they are siblings. But the Bourne shell forks and then the child forks again, so the first process is a child of the second process. Thus when the first process (cat in this case) terminates, the second process (ec in this case) gets a SIGCHLD signal. 

This reveals a signal handling initializing problem in the runtime. When immediately after the signal handler installation, there is a chance that the app receives a signal that needs to be handled before the initialization of the runtime data structures. This is still not fixed by this patch. However we now default all signals to be ignored before the signal handler installation, in order to avoid incorrect handling for signals that are intended to be ignored in between of signal handler installation and the setup for sig_ign[] flags.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@80110 8089f293-4706-0410-a29e-feb5c42a2edf