Fixed a bug in the initial resizing when using `utf_32_substring_to_utf_16_0_pointer' or `utf_32_substring_to_utf_16_pointer (we forgot that UTF-16 code are 2-byte long.
Inlined the content of `utf_32_substring_to_utf_16_pointer' in `utf_32_substring_to_utf_16_0_pointer' and improved the code so that it is exactly the same speed as the original code from WEL_STRING.set_substring. The only difference is that we only perform a resizing when we hit a code greater than 0xFFFF.

Also in `utf_32_substring_to_utf_16_0_pointer' we shrink the result only when resizing the managed pointer, not all the time as this could have an effect on callers who might have allocated a large MANAGED_POINTER and don't want the content to be shrinked since it is null terminated.

See review#8199166.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@89314 8089f293-4706-0410-a29e-feb5c42a2edf