Removed bad conversion trying to write UTF-8 values into .ecf. This fixes the bug Unicode was not correctly retrieved back from .ecf.
Added precondition in `add_attribute' as we do not support Unicode attribute
names. This means that in EIS tool, for parameter pairs like "bookmark=foo", we
only support Unicode values "foo", not the name "bookmark". This is fine at the
moment, as we don't have Unicode attribute names. And the user is unlikely to
put useless parameters with Unicode names. (Only for targets and clusters that
needs .ecf, Eiffel class is fine)

review#9075011

git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@90391 8089f293-4706-0410-a29e-feb5c42a2edf