Added a wrapper class for immutable data structures.
The class can wrap anything conforming to READABLE_INDEXABLE (in particular both LIST and ARRAY objects)
and provides read-only access. It also has some nice features like for_all or do_all and can be used
with the across syntax.

The backend interface has been adapted to use the new class for specifying which attributes to load.
Previously this has been a LIST[STRING] which was dangerous, as backends and plugins sometimes change it
and TYPE_METADATA doesn't like its internal data structures messed up.
With the new class an existing array can be safely wrapped, avoiding the need for an expensive copy of a 
full datastructure when not needed. Furthermore the projection ARRAY from tuple queries doesn't need to be
converted to a list any more.

Furthermore the PLUGIN interface has been changed to take the new immutable strings into account. Instead
of modifying arguments directly during {PS_PLUGIN}.before_retrieve, it can provide new arguments to
replace the old ones. This makes it both safer with respect to aliasing and more powerful.




git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93115 8089f293-4706-0410-a29e-feb5c42a2edf