Added a workaround to support HASH_TABLE.
It turned out that the error was actually within STRING, because
the internal_hash_code doesn't get stored. It is debatable if this
is actually necessary, because a user never gets to see the field
and it can always be generated from the string itself (it merely
serves as a cache).
However, the is_deep_equal feature will not recognize the objects 
as equals, thus breaking the test case. Therefore I've added a 
workaround to generate the hash codes automatically when ABEL 
detects a HASH_TABLE[ANY, READABLE_STRING_GENERAL] object.

This commit fixes the last (known) framework error. The in-memory
repository can now pass all tests, and other tests usually fail 
because collections are not supported.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93138 8089f293-4706-0410-a29e-feb5c42a2edf