Added an object graph traversal algorithm.
The algorithm is defined in PS_OBJECT_GRAPH_TRAVERSAL, which in
turn makes use of the facilities in EiffelBase.
The algorithm basically generates an array of PS_OBJECT_DATA, 
whereas each array element contains a REFLECTED_OBJECT
pointing to the original object, a PS_TYPE_METADATA describing
its type, and a list of all referers and references found
for the object.

The algorithm comes with a small unit test.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93448 8089f293-4706-0410-a29e-feb5c42a2edf