Improved the way additional attributes get handled.
For tuple queries it is important that the backend only
loads the attributes defined in the projection array.
Otherwise all possible performance gain is lost...

Up to now a dedicated PS_PLUGIN was used to remove
superfluous attributes. The drawbacks of this approach
was that there was an overhead for normal queries as
well (where additional attributes aren't a problem)
and the initialization was tricky, because e.g.
the ESCHER plugin depends on the added `version'
attribute (and besides it never actually worked...).

The new mechanism (implemented in QUERY_CURSOR and 
BACKEND_OBJECT) only filters objects for tuple
queries, does not interfere with other plugins,
and should in theory be faster due to using a hash
table for attribute lookup.



git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@93631 8089f293-4706-0410-a29e-feb5c42a2edf