Added initial version for a framework for performance tests.
The main part of the new framework is the EXECUTOR containing
some backend-agnostic tests and an EMITTER, which can 
stop the time and generate output (as a CSV file or stdout).

The idea is to later have different CONTROLLER classes which
create specific backends and can be copy-pasted automatically
by a shell script, before compiling in finalized mode and
starting the test.




git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93678 8089f293-4706-0410-a29e-feb5c42a2edf