Fixed another serious performance issue:
READ_MANAGER.process_next iterated over the set to_process_next.
The initial assumption was that this set would be small all the time.
This is not true any more however since ABEL learned to process a batch
of objects at once. This commit fixes the issue by using the cache
inside the READ_MANAGER.




git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93691 8089f293-4706-0410-a29e-feb5c42a2edf