Changed error handling in the API:
Now if an error happens either in PS_TRANSACTION or 
PS_QUERY, their attributes `has_error' is set to true.
The exception raised by the ABEL backend is no longer
propagated to the user. Instead, every function has
a precondition `not has_error'.

This change triggered a lot of contract and control flow
changes, thus a regression is likely. The unit and
performance tests however pass, and the tutorial works
as long as there is no error.

There's some further work necessary, e.g.  adapting the
tutorial to the changed semantics or checking exception 
propagation and rescue statements.

Also fixed a bug in the MySQL backend where connections
do not get closed correctly.




git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93832 8089f293-4706-0410-a29e-feb5c42a2edf