Improved in-memory backend by using type as a hash key. Replaced some outdated comments by <Precursor> and deleted some dead code.
Removed a double initialization of the PS_AGENT_CRITERION_REMOVER plugin
in the SQL backend and made it remember the plugin list during `wipe_out'.
Also replaced most header comments with <Precursor> and checked identation
for code comments.

Searched for comments and variables which still refer to PS_BACKEND
instead of REPOSITORY_CONNECTOR and renamed them.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@93884 8089f293-4706-0410-a29e-feb5c42a2edf