Fixed an issue when popping elements of a stack by chunks. If the size of number of items to remove is actually greater than the address of the top of the stack then the optimization to figure out if we are still in the first chunk is not working (since it is an unsigned comparison, we get a very high value instead of a value lower than arena) because a comparison becomes true instead of being false. The fix is to make sure we can substract to `top' that size before comparing it.
This was found by running a system on Solaris x86 where pointer addresses where still quite low in the 32-bit address range and the number of items to pop after a retrieve was about 3 millions (i.e. 27Mbytes) and the address was 0x15E3600 (which is only 22Mbytes) so we were getting back a value of 0xFFFFFFFFFFF00000, which was greater than arena. This mistakes caused a crash later because the GC was not protecting all the objects but was still traversing them and we would get a segmentation fault.

There is unfortunately no eweasel test that can reproduce it since we cannot generate the addresses of pointers where we want them to be.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@94572 8089f293-4706-0410-a29e-feb5c42a2edf