Use `const' for most C arrays we generate. Updated most routines to take as arguments `const ...' to show they do not modify the argument. This is necessary due to the first change for C generated arrays.
In `rqst_idrs.c', there was in `idr_Dumped' some code that could now write to `const char *' in the case the type was not EX_CALL, EX_RESC or EX_RETY. But I could not find any situation where we would hit that code. So I removed it. In the event it happens again, we would have to redesign the type of the structure used to communicate.

git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@95858 8089f293-4706-0410-a29e-feb5c42a2edf