Changed agent creation rules for separate targets.
This change makes it possible to introduce library
features that help users to avoid writing small
little wrapper features in SCOOP to control an object.

The patch weakens the rule for agent creation such
that it is now possible to create agents on separate,
uncontrolled target. The newly created agent object
is placed on the same processor as the target, and it
inherits the 'is_controlled' status.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@96546 8089f293-4706-0410-a29e-feb5c42a2edf