Make all separate callbacks synchronous.
This commit makes all separate callbacks synchronous. This
is in line with Benjamin Morandis specification.

It also fixes a bug in the runtime where the call stack lock
and all passed request queue locks of a processor may be
duplicated, meaning that one may end up with race conditions
as two concurrent processors use the same queue on a third one.

Furthermore, it fixes the definition of eif_is_uncontrolled to
include the subordinate status (i.e. if a supplier has
passed its locks). This avoids some unnecessary locking in the
case of separate callbacks and circumvents the bug
described in test#scoop052.

The patch will break test#scoop005, because the order of
execution will change.

See also review#4565990897614848.




git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@96808 8089f293-4706-0410-a29e-feb5c42a2edf