Replace class notify_token with a condition variable.
The commit replaces the C++ notify_token class (used
for SCOOP wait conditions) with a condition variable.
The implementation in C therefore becomes a lot simpler,
because there's no need any more for shared_ptr and
the mpscq class.

The patch changes some of the semantics as well however.
Instead of registering for notifications on all suppliers
within a request group, the client only registers with
those it is currently synchronized with (which means it
has previously sent a query as part of the precondition).
The effects should not be noticeable however on
preconditions that have no side effects.

This change breaks test#scoop012 even further because of
the side effect in the precondition, but at least
it consistently deadlocks now.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@96931 8089f293-4706-0410-a29e-feb5c42a2edf