Added an argument in rt_scoop_setup which indicates whether the project is compiled with SCOOP concurrency settings, and removed the feature init_scoop_root_thread().
Also reverted the initialization back to the old behaviour
where the processor_registry and the root processor are
always created, because of a bug in the interpreter where
the macro RTS_OU is sometimes called in non-SCOOP projects.
This fixes the regression for test#exec184 and test#exec191.

See also review#5522347037556736.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@97112 8089f293-4706-0410-a29e-feb5c42a2edf