Changed the way a terminating SCOOP thread is marked.
Instead of setting the is_processor flag to false, the
logical_id is now set to NULL_PROCESSOR_ID. That way
the garbage collector can distinguish between a regular
thread and a SCOOP processor which is about to be destroyed.

This commit fixes test#scoop028, test#scoop029 and test#bench003.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@97202 8089f293-4706-0410-a29e-feb5c42a2edf