Runtime support for passive processors.
This patch adds rudimentary support for passive processors
in the runtime. It still requires a thread in the background,
which is used to 'lock' a region by sending a special message.

Sometimes a passive processor also has to execute calls on
its own. This is because the SCOOP runtime cannot safely
impersonate a region if not both, client and supplier, support
impersonation (the main reason for this limitation are separate
callbacks). Currently there is no easy fix for this problem,
but in the future we may be able to solve it when there's a
clear distinction between processor ID and region ID.

The commit doesn't break existing eweasel tests, but the new
mechanism is completely untested.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@97581 8089f293-4706-0410-a29e-feb5c42a2edf