Added two global variables for SCOOP.
The new variables are used to keep track of
the SCOOP processor and region ID of a thread.
This information needs to be avaliable for
a faster impersonation mechanism that no longer
needs a setjmp() target.
The variables are introduced in the public
context because in the future they'll be
accessed often by the SCOOP compiler macros.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@97929 8089f293-4706-0410-a29e-feb5c42a2edf