Added a special WAIT_CONDITION_UNLOCK unlock operation.
Previously, whenever a wait condition failed the
client sent a regular unlock message, causing the
supplier to notify all subscribers. The result is
that when two clients were waiting on the same condition,
they both woke each other up in some sort of ping-pong
behaviour.
The new unlock operation does not trigger a notification
to subscribers any more, which prevents this behaviour.

A potential problem of this change is that wait
conditions with side effects are even more broken
(test#scoop009 for example, before rev#98078), but
we don't support them anyway.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@98079 8089f293-4706-0410-a29e-feb5c42a2edf