Basic support for Solaris SPARC with SCOOP.
The patch introduces a new structure rt_atomic_int
which models an atomic integer but will fall back
to locks if there are no atomics on the target
platform.
The attribute 'processor_count' in the processor
registry has been converted to use the new structure.
This is the only case where correctness of atomic
operations is really critical.

Furthermore, the declaration of the rt_processor_registry
struct has been moved to the C file to support
information hiding.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@98207 8089f293-4706-0410-a29e-feb5c42a2edf