Fixed improper type of like argument in generics.
Fixed 14.05 -> 15.08 regression in handling of anchored to argument for void-safety which caused something that should be attached to be detachable because the type of an object does not carry the attachment status and the new code relies on it to create the new type. This is not a problem in experimental mode where the type of an object is always attached. Fixes eweasel test#exec364.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@98251 8089f293-4706-0410-a29e-feb5c42a2edf