Fix invalid implementation of `before`
To provide a proper definition of `before` we actually use indices in the range 1..count and only when we query the pixel do we substract -1. Remove the need of column_value and row_value. Update code accordingly to the above change.

The only breaking change is `index` which is now returning +1 compared to before.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@99025 8089f293-4706-0410-a29e-feb5c42a2edf