Fix stack overflow in test#incr433
When a basic type is defining `item` to be of type `like Current` we were getting a stack overflow when checking the validity of types at the end of degree 2 when calling `{TYPE_A}.is_runnable_on_processor`. This was due to an invalid skeleton generated for `like Current` attributes when we now the type cannot change (case of expanded without loose generics). We fix this by defining `is_loose` on LIKE_CURRENT properly and changed the definition of `description`.

Refactor the `description` and `description_with_detachable_type` to be more uniform and using a new creation procedure to create the instances of ATTR_DESC.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@99048 8089f293-4706-0410-a29e-feb5c42a2edf