Fix invalid default colors for grid selection
When using a text field for the selection, we get invalid colors back from GTK (black on black for selection). Instead of using a text field, we are now using a tree which is actually closer to a grid than a text field with respect to selections. Now we get a nice bluish color and white text.

git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@99075 8089f293-4706-0410-a29e-feb5c42a2edf