Better context saving management
Each draw operations should be surrounded by {CAIRO}.save/{CAIRO}.restore. This way one can change the pencil color between draw operation, not doing so would use the last pencil color for all previous draws.
Better saving of context for an EV_DRAWING_AREA when drawing outside of an expose action, since a cairo context has to be created for each call, it means that we need to restore all set settings that we performed before.

git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@99082 8089f293-4706-0410-a29e-feb5c42a2edf