Change test behavior as it would block on macOS
Instead of just connecting and reading nothing from the socket,
we force the server to send some data and the client to wait until
there is some data to be read.
This was necessary on macOS 10.12 (sierra) as even after connecting
successfully to the socket, the server would hang in the call to `accept'.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@99258 8089f293-4706-0410-a29e-feb5c42a2edf