Handle retrieval of object graph containing old routine objects
When we removed the first formal generic parameter in the routine classes (ROUTINE, PROCEDURE, FUNCTION, PREDICATE) in version 15.12 of EiffelStudio we did a lot of work to ensure that old code could still be compiled without modification. However we forgot to handle the case of storables. This commit addresses this by ignoring the first actual generic parameter of those types from the retrieval system when there is a mismatch on the number of generic parameters of those class.

List of actual changes:
- Moved eif_gen_id_from_cid from gen_conf.c into rt_mapped_cid in retrieve.c
- Added `formal_map` to type_descriptor that map the old formal generic parameter of the routine classes to the new one (here the mapping is straightforward: new_formal_pos = old_formal_pos - 1. However this can be used in the future when handling more complex situation where formals are added/removed/moved.
- Add `is_old_routine_type` to type_descriptor which is just for the purpose of retrieving routine instances (a temporary fix until we have a more general solution to the addition/removal/move of formal generic parameters.)
- Skip/ignore the first actual generic parameter anywhere it was required.



git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@100264 8089f293-4706-0410-a29e-feb5c42a2edf