Fix incorrect signature of veto_pebble_function
Vision2 only accepts a veto pebble function with one argument not two, and the docking library requires
sometime two. Made sure that we pass the right pebble function otherwise `valid_operands` will now yield
false in the internal of Vision2 (before it worked due to our invalid usage of `empty_operands` which
creates the proper tuple to call the function albeit not void safe since the extra argument was Void.


git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@100909 8089f293-4706-0410-a29e-feb5c42a2edf