Added {EB_PROJECT_SESSION_STATISTICS} to track the number of compilations, successful compilations, failed compilations, etc. Added preferences {EB_DEVELOPMENT_WINDOW_PREFERENCES} to Interface->Development Window
Updated the `Src\Delivery\studio\eifinit\default.xml file. with default values for the new prefereces.
pretty_printer_messindex: the difference between the current code and pretty printed code. (Eiffel Style)
consecutive_successful_compilations_threshold: number of successful compilations threshold in a row,
criteria we use to suggest a specific Notification to the user, for example, pretty printer his code.


Updated feature {EB_EDITOR}.prettify, to reset the number of successful compilations in a row that's used by the `prettify notification`.
Added features to {EB_DEVELOPMENT_WINDOW_STATUS_BAR} `on_project_update_statistics`, `pretty_printer`, `editor_prettify` to evaluate when we need to display a notification Prettify Class.

The current code shows a Notification with an `Apply` option and the other one with a message letting the user know-how
to use the short-cut to apply the pretty printer.


git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@104605 8089f293-4706-0410-a29e-feb5c42a2edf