feat: Added the cloud user profile functionality.
For now basic, just include a wikitext "about" field, so that user can describe themselves, add projects information, and so on.
Added SHOP_CUSTOMER, to get internal information of a customer, mostly from stripe if this is used as payment system.
Updated various cloud views, to have link on that new user profile page.

git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@105109 8089f293-4706-0410-a29e-feb5c42a2edf